Updated README.md
This Model helps evaluate whether applicants will be successful if funded by Alphabet Soup.
By using knowledge with machine learning and neural networks, and decide to use the features in the provided dataset to create a binary classifier model that will predict whether an applicant will become a successful business.

---
## Technologies

- [Anaconda](https://www.anaconda.com/products/individual) - Pandas is included in Anaconda distribution and Conda package manager to manage Python environments.
- [Jupyter Lab](https://jupyter.org/) - web-based user interface designed for data analysis. It lets you write, run, and review the results in Python programs (all in a single integrated development environment (IDE).
- [The imbalanced-learn documentation](https://imbalanced-learn.org/stable) - mbalanced-learn (imported as imblearn) is an open source, MIT-licensed library relying on scikit-learn (imported as sklearn) and provides tools when dealing with classification with imbalanced classes.
 - [The scikit-learn documentation](https://scikit-learn.org/stable) - Simple and efficient tools for predictive data analysis.
 - [sklearn.preprocessing.OneHotEncoder]( https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.OneHotEncoder.html)
 - [TensorFlow 2.0]( https://www.tensorflow.org) – An end-to-end open source machine learning platform.
 - [Keras]( https://keras.io/about) - Keras is a deep learning API written in Python, running on top of the machine learning platform TensorFlow. It was developed with a focus on enabling fast experimentation. Being able to go from idea to result as fast as possible is key to doing good research.

---
## Installation Guide

To install TensorFlow, open the terminal, and execute the following command:
Use the pip install command to install the TensorFlow 2.0 library.
```
Pip install –upgrade tensorflow
```
Verify Installation
Once the TensorFlow install is complete, verify that the installation completed successfully.
```
python -c ”import tensorflow as tf;print(tf._version_)”
```
The output of this command should show version 2.0.0 or higher.

Keras

Keras is a popular deep learning framework that serves as a high-level API for TensorFlow. Keras is now included with TensorFlow 2.0. So, run the following command to verify that the package is available:
```
python -c ”import tensorflow as tf;print(tf.keras._version_)”
```
The output should show version 2.2.4-tf or later.

Apple M1 Chip Users
The Apple M1 Chip is not currently compatible with a typical Tensorflow installation, but that's ok! You can easily run all of the activities in this module using Google Colab (Links to an external site.). Each activity will feature a notebook file that is streamlined to run smoothly in Google Colab. To complete the activities in this module, visit Google Colab (Links to an external site.), upload the notebook file for the activity that is named with a "GC" (Google Colab) prefix and run the code as you would normally.


## Contributors

Created by Bina Jariwala

---

## License

None

---